NhanUQ update code BE

+++ C0722G1-Repo-BE/src/main/java/com/c0722g1repobe/service/employee/impl/EmployeeService.java
@@ -63,10 +63,16 @@ public class EmployeeService implements IEmployeeService {
      * Function: delete employee
      *
      * @param id
+     * @return true if delete success or false if id not exist
      */
     @Override
-    public void deleteEmployee(Long id) {
+    public boolean isDeleteEmployee(Long id) {
+        Optional<Employee> employee = employeeRepository.getByIdEmployee(id);
+        if (!employee.isPresent()) {
+            return false;
+        }
         employeeRepository.deleteEmployee(id);
+        return true;
     }
 
     /**
